implement some of the unreleased changes from master
excluded https://github.com/cloudtools/stacker/pull/564 for now
