+++ tests/cfngin/test_util.py
@@ -1,24 +1,20 @@
 """Tests for runway.cfngin.util."""
-import os
-import queue
-import string
 # pylint: disable=unused-argument,invalid-name
+import os
+import string
 import unittest
 
 import boto3
 import mock
 
-from runway.cfngin.config import GitPackageSource, Hook
+from runway.cfngin.config import GitPackageSource
 from runway.cfngin.util import (Extractor, SourceProcessor, TarExtractor,
                                 TarGzipExtractor, ZipExtractor, camel_to_snake,
                                 cf_safe_name, get_client_region,
-                                get_s3_endpoint, handle_hooks,
-                                load_object_from_string, merge_map,
-                                parse_cloudformation_template,
+                                get_s3_endpoint, load_object_from_string,
+                                merge_map, parse_cloudformation_template,
                                 s3_bucket_location_constraint,
                                 yaml_to_ordered_dict)
-
-from .factories import mock_context, mock_provider
 
 AWS_REGIONS = ["us-east-1", "cn-north-1", "ap-northeast-1", "eu-west-1",
                "ap-southeast-1", "ap-southeast-2", "us-west-2", "us-gov-west-1",
@@ -270,167 +266,6 @@ Outputs:
                 sp.determine_git_ref({'uri': 'git@foo', 'tag': 'v1.0.0'}),
                 'v1.0.0'
             )
-
-
-hook_queue = queue.Queue()
-
-
-def mock_hook(*args, **kwargs):
-    """Mock hook."""
-    hook_queue.put(kwargs)
-    return True
-
-
-def fail_hook(*args, **kwargs):
-    """Fail hook."""
-    return None
-
-
-def exception_hook(*args, **kwargs):
-    """Exception hook."""
-    raise Exception
-
-
-def context_hook(*args, **kwargs):
-    """Context hook."""
-    return "context" in kwargs
-
-
-def result_hook(*args, **kwargs):
-    """Results hook."""
-    return {"foo": "bar"}
-
-
-class TestHooks(unittest.TestCase):
-    """Tests for runway.cfngin.util hooks."""
-
-    def setUp(self):
-        """Run before tests."""
-        self.context = mock_context(namespace="namespace")
-        self.provider = mock_provider(region="us-east-1")
-
-    def test_empty_hook_stage(self):
-        """Test empty hook stage."""
-        hooks = []
-        handle_hooks("fake", hooks, self.provider, self.context)
-        self.assertTrue(hook_queue.empty())
-
-    def test_missing_required_hook(self):
-        """Test missing required hook."""
-        hooks = [Hook({"path": "not.a.real.path", "required": True})]
-        with self.assertRaises(ImportError):
-            handle_hooks("missing", hooks, self.provider, self.context)
-
-    def test_missing_required_hook_method(self):
-        """Test missing required hook method."""
-        hooks = [{"path": "runway.cfngin.hooks.blah", "required": True}]
-        with self.assertRaises(AttributeError):
-            handle_hooks("missing", hooks, self.provider, self.context)
-
-    def test_missing_non_required_hook_method(self):
-        """Test missing non required hook method."""
-        hooks = [Hook({"path": "runway.cfngin.hooks.blah", "required": False})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        self.assertTrue(hook_queue.empty())
-
-    def test_default_required_hook(self):
-        """Test default required hook."""
-        hooks = [Hook({"path": "runway.cfngin.hooks.blah"})]
-        with self.assertRaises(AttributeError):
-            handle_hooks("missing", hooks, self.provider, self.context)
-
-    def test_valid_hook(self):
-        """Test valid hook."""
-        hooks = [
-            Hook({"path": "tests.cfngin.test_util.mock_hook",
-                  "required": True})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        good = hook_queue.get_nowait()
-        self.assertEqual(good["provider"].region, "us-east-1")
-        with self.assertRaises(queue.Empty):
-            hook_queue.get_nowait()
-
-    def test_valid_enabled_hook(self):
-        """Test valid enabled hook."""
-        hooks = [
-            Hook({"path": "tests.cfngin.test_util.mock_hook",
-                  "required": True, "enabled": True})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        good = hook_queue.get_nowait()
-        self.assertEqual(good["provider"].region, "us-east-1")
-        with self.assertRaises(queue.Empty):
-            hook_queue.get_nowait()
-
-    def test_valid_enabled_false_hook(self):
-        """Test valid enabled false hook."""
-        hooks = [
-            Hook({"path": "tests.cfngin.test_util.mock_hook",
-                  "required": True, "enabled": False})]
-        handle_hooks("missing", hooks, self.provider, self.context)
-        self.assertTrue(hook_queue.empty())
-
-    def test_context_provided_to_hook(self):
-        """Test context provided to hook."""
-        hooks = [
-            Hook({"path": "tests.cfngin.test_util.context_hook",
-                  "required": True})]
-        handle_hooks("missing", hooks, "us-east-1", self.context)
-
-    def test_hook_failure(self):
-        """Test hook failure."""
-        hooks = [
-            Hook({"path": "tests.cfngin.test_util.fail_hook",
-                  "required": True})]
-        with self.assertRaises(SystemExit):
-            handle_hooks("fail", hooks, self.provider, self.context)
-        hooks = [{"path": "tests.cfngin.test_util.exception_hook",
-                  "required": True}]
-        with self.assertRaises(Exception):
-            handle_hooks("fail", hooks, self.provider, self.context)
-        hooks = [
-            Hook({"path": "tests.cfngin.test_util.exception_hook",
-                  "required": False})]
-        # Should pass
-        handle_hooks("ignore_exception", hooks, self.provider, self.context)
-
-    def test_return_data_hook(self):
-        """Test return data hook."""
-        hooks = [
-            Hook({
-                "path": "tests.cfngin.test_util.result_hook",
-                "data_key": "my_hook_results"
-            }),
-            # Shouldn't return data
-            Hook({
-                "path": "tests.cfngin.test_util.context_hook"
-            })
-        ]
-        handle_hooks("result", hooks, "us-east-1", self.context)
-
-        self.assertEqual(
-            self.context.hook_data["my_hook_results"]["foo"],
-            "bar"
-        )
-        # Verify only the first hook resulted in stored data
-        self.assertEqual(
-            list(self.context.hook_data.keys()), ["my_hook_results"]
-        )
-
-    def test_return_data_hook_duplicate_key(self):
-        """Test return data hook duplicate key."""
-        hooks = [
-            Hook({
-                "path": "tests.cfngin.test_util.result_hook",
-                "data_key": "my_hook_results"
-            }),
-            Hook({
-                "path": "tests.cfngin.test_util.result_hook",
-                "data_key": "my_hook_results"
-            })
-        ]
-
-        with self.assertRaises(KeyError):
-            handle_hooks("result", hooks, "us-east-1", self.context)
 
 
 class MockException1(Exception):
